:recycle: Code refactoring, :fire: Removed app bar, :sparkles: Added Bottom App Bar
