+++ lib/values/assets.dart
@@ -1,0 +1,3 @@
+class Assets {
+  static const String featureGraphic = 'assets/images/feature-graphic.png';
+}
