+++ lib/values/imports.dart
@@ -11,6 +11,7 @@ export 'package:flutter_playground/ui/animatedicons/AnimatedIcons.dart';
 export 'package:flutter_playground/ui/animatedsize/AnimatedSize.dart';
 export 'package:flutter_playground/ui/animatedswitcher/AnimatedSwitcherExample.dart';
 export 'package:flutter_playground/ui/app_bar/AppBarExample.dart';
+export 'package:flutter_playground/ui/app_bar_bottom/appbar_bottom.dart';
 export 'package:flutter_playground/ui/backdrop_filter/BackdropFilterExample.dart';
 export 'package:flutter_playground/ui/bottomnavigation/BottomNavigation.dart';
 export 'package:flutter_playground/ui/chipsexample/ChipsExample.dart';
@@ -18,10 +19,13 @@ export 'package:flutter_playground/ui/collapsibletoolbar/CollapsibleToolbar.dart
 export 'package:flutter_playground/ui/cupertino/CupertinoActionSheetExample.dart';
 export 'package:flutter_playground/ui/cupertino/CupertinoProgressIndicatorExample.dart';
 export 'package:flutter_playground/ui/cupertino/CupertinoTimerPickerExample.dart';
+export 'package:flutter_playground/ui/datatable/data_table_example.dart';
 export 'package:flutter_playground/ui/dismissible/DismissibleExample.dart';
 export 'package:flutter_playground/ui/dragdrop/ExampleDragDrop.dart';
 export 'package:flutter_playground/ui/drawer/NavigationDrawer.dart';
+export 'package:flutter_playground/ui/expanded_example/expanded_example.dart';
 export 'package:flutter_playground/ui/expansion_tile/ExpansionTileExample.dart';
+export 'package:flutter_playground/ui/flare/flare_example.dart';
 export 'package:flutter_playground/ui/flow_example/FlowWidgetExample.dart';
 export 'package:flutter_playground/ui/gridpaper/GridPaperExample.dart';
 export 'package:flutter_playground/ui/hardwarekey/RawKeyboardDemo.dart';
@@ -30,6 +34,7 @@ export 'package:flutter_playground/ui/local_auth/LocalAuth.dart';
 export 'package:flutter_playground/ui/map/GoogleMapsExample.dart';
 export 'package:flutter_playground/ui/nestedlist/NestedList.dart';
 export 'package:flutter_playground/ui/progressbutton/ProgressButton.dart';
+export 'package:flutter_playground/ui/quick_actions/QuickActionsExample.dart';
 export 'package:flutter_playground/ui/rotatedbox/RotatedBox.dart';
 export 'package:flutter_playground/ui/rotation_transition/RotationTransitionExample.dart';
 export 'package:flutter_playground/ui/staggeredanimation/StaggerDemo.dart';
@@ -39,4 +44,6 @@ export 'package:flutter_playground/ui/text/TextExamples.dart';
 export 'package:flutter_playground/ui/text/TextSpan.dart';
 export 'package:flutter_playground/ui/text/TextUnderline.dart';
 export 'package:flutter_playground/ui/tooltip/tooltip.dart';
+export 'package:flutter_playground/ui/transform/transform_example.dart';
+export 'package:flutter_playground/ui/wrap/wrap_example.dart';
 export 'package:flutter_playground/utils/Strings.dart';
