+++ pubspec.yaml
@@ -11,7 +11,7 @@ description: A Flutter application contains a list of Flutter Examples.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.2.3+6
+version: 1.2.4+7
 
 environment:
   sdk: ">=2.7.0-dev.2.1 <3.0.0"
